add random spawn mobs on map

+++ src/bin/deleteAllTiles.ts
@@ -1,15 +1,17 @@
 import { prisma } from '../utils/prisma';
 
-export const deleteAllTiles = async () => {
-  const allTiles = (await prisma.tile.findMany({}))
-    .filter((item) => item.objectId)
-    .map((item) => item.objectId);
-  for (const tile of allTiles) {
-    await prisma.tile.deleteMany({
-      where: {
-        objectId: tile,
+export const newDeleteTiles = async () => {
+  const allTiles = await prisma.tile.findMany({});
+  const filteredTiles = allTiles.filter((tile) => tile.objectId);
+
+  if (!filteredTiles.length) return;
+
+  await prisma.tile.deleteMany({
+    where: {
+      objectId: {
+        in: filteredTiles.map((tile) => tile.objectId as string),
       },
-    });
-  }
+    },
+  });
   await prisma.tile.deleteMany();
 };
